complete separation of ANTLR module
