+++ antlr/src/main/kotlin/com/strumenta/kolasu/parsing/FirstStageParsingResult.kt
