+++ core/src/main/kotlin/com/strumenta/kolasu/language/LanguageModule.kt
@@ -2,8 +2,9 @@ package com.strumenta.kolasu.language
 
 import com.strumenta.kolasu.codegen.ASTCodeGenerator
 import com.strumenta.kolasu.model.Node
+import com.strumenta.kolasu.parsing.ASTParser
 
 /**
  * This permits to parse code into AST and viceversa going from an AST into code.
  */
-class LanguageModule<R : Node>(val parser: KolasuParser<R, *, *, *>, val codeGenerator: ASTCodeGenerator<R>)
+class LanguageModule<R : Node>(val parser: ASTParser<R>, val codeGenerator: ASTCodeGenerator<R>)
